State no clients authorized instead of no display.
Previously if no clients were authorized - even though this is clearly a
valid situation - the user would display a blank screen. To be
consistent with the other sibling Vue components this proposal is to
state that no clients are authorized.

This has the following benefits:

* It's obvious the Vue component is working!

* If someone is using this, authorizes a client, they don't have to
  figure out if the Vue component is working OR if they've done
  something wrong in their testing or setup.

The test may not be elegant however testing for tokens.length === 0
didn't catch when no clients at all had been authorized. If there is a
more elegant (or more Vue like way), that would be really good.