Accept * scope as valid when verifying client_credentials
When using the middleware 'client_credentials:some_scope', requests are rejected if the
token was requested using '*' scope during a client credentials grant. They are accepted
if the token was explicitly requested using 'some_scope', in this case.

If I understand correctly that the '*' scope, if granted, would allow all requests for
routes protected by any scope, then this pull request should allow this to happen.

Tests updated to reflect the change.