Token Issuing Should Use Api Middleware Group
Currently, the token issuing route only uses the throttle middleware, but it should use all the middlewares defined in the "api" middleware group, giving more control over this to the developer. This will allow to the programmer to include additional middleware to this route quickly.

For example, CORS middleware, of course, the throttling, and other custom middleware.

As it is working currently, the developer must register this route again in the routes/api.php to use all the "api" middleware.