Remove decrypt from TokenGuard.
After following the instructions for "Consuming Your API with JavaScript" the api requests return 401 unauthorized. This was because the JWT cookie being added to the site was not encrypted using laravels encryption service, however when it was being processed in the TokenGuard it was decrypted using laravels encryption service which threw an error.
This could either be fixed by encrypting the cookie before it is sent or removing the decryption code when it is received. I chose the latter as the token is encoded in the php-jwt package.
I apologise if I have missed something and this is the correct behaviour, or if I did the fix the wrong way round.