Use unsignedInteger column type for client_id columns
This is because the client id autoincrements. We don't add foreign key indexes ourselves but if you try to add them afterwards you'll have to first change the column type of the client_id columns because otherwise they don't match the increments column. We can immediately use the proper type in the migrations here.