Fix revoked refresh token check
When a check if performed to determine if a refresh token is revoked or not, it's unnecessary to also check if its corresponding access token is revoked. It's perfectly possible to issue a new access token if the old access token is either revoked or pruned as long as the refresh token is valid.

Fixes https://github.com/laravel/passport/issues/917