REFACTOR: Standardize all routes with the pattern `passport.*.*.<action>`
The route `passport.tokens` was the only one not complying with the naming pattern adopted on the rest of the project's routes. Also it can mislead to the interpretation that if the name omitted the action, it might be a getter.

In order to let this organized in the same standard, I am suggesting this refactor.